+++ ClubManager/QueryObjects/NewStudentQO.cs
@@ -4,7 +4,7 @@ namespace ClubManager.QueryObjects
     {
         public string Number { get; set; }
         public string Name { get; set; }
-        public byte? Grade { get; set; }
+        public long? Grade { get; set; }
         public string Major { get; set; }
         public string Phone { get; set; }
     }
